Version 2.0
Changes from old revision

The RGB resistors provided are 90% smaller in value, silkscreen indication has been fixed. The filtered output signal is brighter, it can be considered an improvement.

The VGA input connector is now also female, to prioritize use in a case and more accessible to the user. However, a Mini Gender Changer is also offered, but this time male-male, to connect directly to the video output or to the monitor input, and can be left permanently.

New PCB design with through hole technology and mixed SMD, for an assembly task adapted to the needs of each one.